smac_planner_hybrid: return integer bin from getAngle to avoid OOR index
Theta close to 2*PI (e.g. from OMPL interpolate returning a slightly
negative angle) produced a double bin just below num_angle_quantization
that rounded up to exactly N when stored in the caller's float at
analytic_expansion.cpp. Downstream that bin indexed
oriented_footprints_[N], one past the end. Empirically reproduced for
every power-of-2 quantization (16/32/64/128/256); default 72 was
incidentally safe due to float rounding.

Return unsigned int directly (matching getClosestAngularBin) so the
< N invariant survives float storage. Also wrap reals[2] with fmod in
analytic_expansion so the exact 2*PI boundary and arbitrary negative
inputs normalize cleanly. Same class of bug as #5501 (Lattice planner).

+++ nav2_smac_planner/include/nav2_smac_planner/node_hybrid.hpp
@@ -96,11 +96,11 @@ struct HybridMotionTable
   float getAngleFromBin(const unsigned int & bin_idx);
 
   /**
-   * @brief Get the angle scaled across bins from a raw orientation
-   * @param theta Angle in radians
-   * @return angle scaled across bins
+   * @brief Get the angle bin index for an orientation in [0, 2*PI)
+   * @param theta Angle in radians; caller must ensure theta >= 0
+   * @return bin index in [0, num_angle_quantization)
    */
-  double getAngle(const double & theta);
+  unsigned int getAngle(const double & theta);
 
   MotionModel motion_model = MotionModel::UNKNOWN;
   MotionPoses projections;
